[Ingestion pipeline] Ingest observations without GroupByKey (GBK)

+++ pipeline/ingestion/src/main/java/org/datacommons/ingestion/pipeline/IngestionPipelineOptions.java
@@ -45,7 +45,7 @@ public interface IngestionPipelineOptions extends PipelineOptions {
   void setImportGroupVersion(String importGroupVersion);
 
   @Description("The number of shards to generate for writing mutations.")
-  @Default.Integer(1000)
+  @Default.Integer(1)
   int getNumShards();
 
   void setNumShards(int numShards);
